requirements.txt
streamlit
networkx
rank_bm25
lark
duckduckgo-search
langchain_experimental==0.0.64
sentence-transformers
scikit-learn
python-dotenv==1.0.1
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.8
langchain-chroma==0.1.2
unstructured==0.14.4
chromadb==0.5.0
openai==1.31.1
tiktoken==0.7.0
langchainhub
tenacity
numpy
transformers==4.41.0
tokenizers==0.19.1
